Primera versión del CRUD en rentas y como primer diseño de regla de negocios para la renta finalizado

+++ database/migrations/2023_04_17_052231_create_t_pagos_rentas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->tinyIncrements('clvPagoRenta');
             $table->decimal('pagoRenta', 10, 2)->nullable();
             $table->decimal('ivaRenta', 10, 2)->nullable();
-            $table->unsignedTinyInteger('clvRenta')->nullable();
+            $table->unsignedBigInteger('clvRenta')->nullable();
             $table->unsignedTinyInteger('clvEstadoPagoRenta')->nullable();
             $table->text('descripcion')->nullable();
             $table->timestamps();
